refactor: move wallpaperengine launcher helpers
