+++ waypaper/common.py
@@ -12,12 +12,12 @@ import hashlib
 from pathlib import Path
 from typing import List
 from PIL import Image
-import json
 
 gi.require_version("Gtk", "3.0")
 from gi.repository import Gtk, GdkPixbuf, Gdk, GLib
 
 from waypaper.options import IMAGE_EXTENSIONS, BACKEND_OPTIONS, VIDEO_EXTENSIONS
+from waypaper.wallpaperengine import get_wallpaperengine_preview
 
 
 def has_image_extension(file_path: str, backend: str) -> bool:
@@ -64,33 +64,6 @@ def get_image_paths(backend: str,
                     continue
                 image_path_list.append(os.path.join(root, image_name))
     return image_path_list
-
-def get_wallpaperengine_preview(wallpaperengine_folder: Path | str) -> List[str]:
-    image_path_list = []
-    for root, directories, files in os.walk(wallpaperengine_folder):
-        for file in files:
-            if Path(file).stem == "preview":
-                image_path_list.append(os.path.join(root, file))
-    return image_path_list
-
-
-def get_wallpaperengine_project(full_path: Path | str) -> dict:
-    full_path = Path(full_path)
-    image_dir = full_path.parent
-    with open(image_dir / "project.json", "r") as f:
-        project = json.load(f)
-    wallpaper_type = str(project.get("type", "unknown")).strip().lower() or "unknown"
-    project["type"] = wallpaper_type
-    project["title"] = str(project.get("title") or image_dir.name)
-    return project
-
-def get_wallpaperengine_image_name(full_path: Path | str) -> str:
-    full_path = Path(full_path)
-    image_dir = full_path.parent
-    with open(image_dir / "project.json", "r") as f:
-        project = json.load(f)
-    return project["title"]
-
 
 def get_image_name(full_path: str, base_folder_list: list[Path], include_path: bool) ->  str | None:
     """Get image name that may or may not include parent folders"""
